move genesis test to chain spec
