+++ crates/katana/executor/tests/fixtures/mod.rs
@@ -7,6 +7,7 @@ use katana_primitives::block::{
     Block, ExecutableBlock, FinalityStatus, GasPrices, PartialHeader, SealedBlockWithStatus,
 };
 use katana_primitives::chain::ChainId;
+use katana_primitives::chain_spec::{self, ChainSpec};
 use katana_primitives::class::{CompiledClass, FlattenedSierraClass};
 use katana_primitives::contract::ContractAddress;
 use katana_primitives::env::{CfgEnv, FeeTokenAddressses};
@@ -48,7 +49,11 @@ pub fn contract_class() -> (CompiledClass, FlattenedSierraClass) {
 
 #[rstest::fixture]
 #[once]
-pub fn genesis() -> Genesis {
+pub fn chain() -> ChainSpec {
+    let mut chain = chain_spec::DEV_UNALLOCATED.clone();
+
+    // to generate the exact list of accounts as you would when you just run `katana` w/o
+    // any additional flags
     let mut seed = [0u8; 32];
     seed[0] = b'0';
 
@@ -57,15 +62,14 @@ pub fn genesis() -> Genesis {
         .with_balance(U256::from(DEFAULT_PREFUNDED_ACCOUNT_BALANCE))
         .generate();
 
-    let mut genesis = Genesis::default();
-    genesis.extend_allocations(accounts.into_iter().map(|(k, v)| (k, v.into())));
-    genesis
+    chain.genesis.extend_allocations(accounts.into_iter().map(|(k, v)| (k, v.into())));
+    chain
 }
 
 /// Returns a state provider with some prefilled states.
 #[rstest::fixture]
-pub fn state_provider(genesis: &Genesis) -> Box<dyn StateProvider> {
-    let states = genesis.state_updates();
+pub fn state_provider(chain: &ChainSpec) -> Box<dyn StateProvider> {
+    let states = chain.state_updates();
     let provider = InMemoryProvider::new();
 
     let block = SealedBlockWithStatus {
