+++ crates/katana/primitives/src/genesis/json.rs
@@ -30,12 +30,9 @@ use super::constant::{
 };
 use super::constant::{
     DEFAULT_ACCOUNT_CLASS, DEFAULT_ACCOUNT_CLASS_CASM, DEFAULT_ACCOUNT_CLASS_HASH,
-    DEFAULT_ACCOUNT_COMPILED_CLASS_HASH, DEFAULT_ETH_FEE_TOKEN_ADDRESS, DEFAULT_LEGACY_ERC20_CASM,
-    DEFAULT_LEGACY_ERC20_CLASS_HASH, DEFAULT_LEGACY_ERC20_COMPILED_CLASS_HASH,
-    DEFAULT_LEGACY_UDC_CASM, DEFAULT_LEGACY_UDC_CLASS_HASH, DEFAULT_LEGACY_UDC_COMPILED_CLASS_HASH,
-    DEFAULT_UDC_ADDRESS,
+    DEFAULT_ACCOUNT_COMPILED_CLASS_HASH,
 };
-use super::{FeeTokenConfig, Genesis, GenesisAllocation, UniversalDeployerConfig};
+use super::{Genesis, GenesisAllocation};
 use crate::block::{BlockHash, BlockNumber, GasPrices};
 use crate::class::{ClassHash, CompiledClass, SierraClass};
 use crate::contract::{ContractAddress, StorageKey, StorageValue};
@@ -146,32 +143,6 @@ impl<'de> Deserialize<'de> for ClassNameOrHash {
 
         deserializer.deserialize_any(_Visitor)
     }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct FeeTokenConfigJson {
-    pub name: String,
-    pub symbol: String,
-    pub address: Option<ContractAddress>,
-    pub decimals: u8,
-    /// The class hash of the fee token contract.
-    /// If not provided, the default fee token class is used.
-    pub class: Option<ClassNameOrHash>,
-    /// To initialize the fee token contract storage
-    pub storage: Option<BTreeMap<StorageKey, StorageValue>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct UniversalDeployerConfigJson {
-    /// The address of the universal deployer contract.
-    /// If not provided, the default UD address is used.
-    pub address: Option<ContractAddress>,
-    /// The class hash of the universal deployer contract.
-    /// If not provided, the default UD class is used.
-    pub class: Option<ClassNameOrHash>,
-    /// To initialize the UD contract storage
-    pub storage: Option<BTreeMap<StorageKey, StorageValue>>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -396,103 +367,6 @@ impl TryFrom<GenesisJson> for Genesis {
             classes.insert(class_hash, GenesisClass { compiled_class_hash, sierra, casm });
         }
 
-        // let fee_token = FeeTokenConfig {
-        //     name: value.fee_token.name,
-        //     symbol: value.fee_token.symbol,
-        //     decimals: value.fee_token.decimals,
-        //     address: value.fee_token.address.unwrap_or(DEFAULT_ETH_FEE_TOKEN_ADDRESS),
-        //     class_hash: match value.fee_token.class {
-        //         Some(ClassNameOrHash::Hash(class_hash)) => class_hash,
-        //         Some(ClassNameOrHash::Name(ref class_name)) => *class_names
-        //             .get(class_name)
-        //             .ok_or_else(|| GenesisJsonError::UnknownClassName(class_name.clone()))?,
-        //         None => DEFAULT_LEGACY_ERC20_CLASS_HASH,
-        //     },
-        //     storage: value.fee_token.storage,
-        // };
-
-        // match value.fee_token.class {
-        //     Some(ClassNameOrHash::Hash(hash)) => {
-        //         if !classes.contains_key(&hash) {
-        //             return Err(GenesisJsonError::MissingClass(hash));
-        //         }
-        //     }
-
-        //     Some(ClassNameOrHash::Name(name)) => {
-        //         let hash = class_names
-        //             .get(&name)
-        //             .ok_or_else(|| GenesisJsonError::UnknownClassName(name.clone()))?;
-
-        //         if !classes.contains_key(hash) {
-        //             return Err(GenesisJsonError::MissingClass(*hash));
-        //         }
-        //     }
-
-        //     // if no class hash is provided, use the default fee token class
-        //     None => {
-        //         let _ = classes.insert(
-        //             DEFAULT_LEGACY_ERC20_CLASS_HASH,
-        //             GenesisClass {
-        //                 sierra: None,
-        //                 casm: Arc::new(DEFAULT_LEGACY_ERC20_CASM.clone()),
-        //                 compiled_class_hash: DEFAULT_LEGACY_ERC20_COMPILED_CLASS_HASH,
-        //             },
-        //         );
-        //     }
-        // };
-
-        // let universal_deployer = if let Some(config) = value.universal_deployer {
-        //     match config.class {
-        //         Some(ClassNameOrHash::Hash(hash)) => {
-        //             if !classes.contains_key(&hash) {
-        //                 return Err(GenesisJsonError::MissingClass(hash));
-        //             }
-
-        //             Some(UniversalDeployerConfig {
-        //                 class_hash: hash,
-        //                 address: config.address.unwrap_or(DEFAULT_UDC_ADDRESS),
-        //                 storage: config.storage,
-        //             })
-        //         }
-
-        //         Some(ClassNameOrHash::Name(name)) => {
-        //             let hash = class_names
-        //                 .get(&name)
-        //                 .ok_or_else(|| GenesisJsonError::UnknownClassName(name))?;
-
-        //             if !classes.contains_key(hash) {
-        //                 return Err(GenesisJsonError::MissingClass(*hash));
-        //             }
-
-        //             Some(UniversalDeployerConfig {
-        //                 class_hash: *hash,
-        //                 address: config.address.unwrap_or(DEFAULT_UDC_ADDRESS),
-        //                 storage: config.storage,
-        //             })
-        //         }
-
-        //         // if no class hash is provided, use the default UD contract parameters
-        //         None => {
-        //             let class_hash = DEFAULT_LEGACY_UDC_CLASS_HASH;
-        //             let address = config.address.unwrap_or(DEFAULT_UDC_ADDRESS);
-        //             let storage = config.storage;
-
-        //             let _ = classes.insert(
-        //                 DEFAULT_LEGACY_UDC_CLASS_HASH,
-        //                 GenesisClass {
-        //                     sierra: None,
-        //                     casm: Arc::new(DEFAULT_LEGACY_UDC_CASM.clone()),
-        //                     compiled_class_hash: DEFAULT_LEGACY_UDC_COMPILED_CLASS_HASH,
-        //                 },
-        //             );
-
-        //             Some(UniversalDeployerConfig { class_hash, address, storage })
-        //         }
-        //     }
-        // } else {
-        //     None
-        // };
-
         let mut allocations: BTreeMap<ContractAddress, GenesisAllocation> = BTreeMap::new();
 
         for (address, account) in value.accounts {
@@ -598,9 +472,7 @@ impl TryFrom<GenesisJson> for Genesis {
 
         Ok(Genesis {
             classes,
-            // fee_token,
             allocations,
-            // universal_deployer,
             number: value.number,
             sequencer_address: value.sequencer_address,
             timestamp: value.timestamp,
@@ -686,26 +558,6 @@ mod tests {
         assert_eq!(json.state_root, felt!("0x99"));
         assert_eq!(json.gas_prices.eth, 1111);
         assert_eq!(json.gas_prices.strk, 2222);
-
-        // assert_eq!(json.fee_token.address, Some(address!("0x55")));
-        // assert_eq!(json.fee_token.name, String::from("ETHER"));
-        // assert_eq!(json.fee_token.symbol, String::from("ETH"));
-        // assert_eq!(json.fee_token.class, Some(ClassNameOrHash::Name(String::from("MyErc20"))));
-        // assert_eq!(json.fee_token.decimals, 18);
-        // assert_eq!(
-        //     json.fee_token.storage,
-        //     Some(BTreeMap::from([(felt!("0x111"), felt!("0x1")), (felt!("0x222"),
-        // felt!("0x2"))])) );
-
-        // assert_eq!(
-        //     json.universal_deployer.clone().unwrap().address,
-        //     Some(address!("0x041a78e741e5af2fec34b695679bc6891742439f7afb8484ecd7766661ad02bf"))
-        // );
-        // assert_eq!(json.universal_deployer.unwrap().class, None);
-        // assert_eq!(
-        //     json.fee_token.storage,
-        //     Some(BTreeMap::from([(felt!("0x111"), felt!("0x1")), (felt!("0x222"),
-        // felt!("0x2")),])) );
 
         let acc_1 = address!("0x66efb28ac62686966ae85095ff3a772e014e7fbf56d4c5f6fac5606d4dde23a");
         let acc_2 = address!("0x6b86e40118f29ebe393a75469b4d926c7a44c2e2681b6d319520b7c1156d114");
@@ -862,32 +714,6 @@ mod tests {
 
         let expected_classes = BTreeMap::from([
             (
-                felt!("0x07b3e05f48f0c69e4a65ce5e076a66271a527aff2c34ce1083ec6e1526997a69"),
-                GenesisClass {
-                    sierra: None,
-                    casm: DEFAULT_LEGACY_UDC_CASM.clone().into(),
-                    compiled_class_hash: felt!(
-                        "0x07b3e05f48f0c69e4a65ce5e076a66271a527aff2c34ce1083ec6e1526997a69"
-                    ),
-                },
-            ),
-            (
-                felt!("0x80085"),
-                GenesisClass {
-                    sierra: None,
-                    casm: DEFAULT_LEGACY_UDC_CASM.clone().into(),
-                    compiled_class_hash: felt!("0x80085"),
-                },
-            ),
-            (
-                felt!("0x8"),
-                GenesisClass {
-                    sierra: None,
-                    compiled_class_hash: felt!("0x8"),
-                    casm: DEFAULT_LEGACY_ERC20_CASM.clone().into(),
-                },
-            ),
-            (
                 DEFAULT_ACCOUNT_CLASS_HASH,
                 GenesisClass {
                     compiled_class_hash: DEFAULT_ACCOUNT_COMPILED_CLASS_HASH,
@@ -905,18 +731,6 @@ mod tests {
                 },
             ),
         ]);
-
-        let expected_fee_token = FeeTokenConfig {
-            address: address!("0x55"),
-            name: String::from("ETHER"),
-            symbol: String::from("ETH"),
-            decimals: 18,
-            class_hash: felt!("0x8"),
-            storage: Some(BTreeMap::from([
-                (felt!("0x111"), felt!("0x1")),
-                (felt!("0x222"), felt!("0x2")),
-            ])),
-        };
 
         let acc_1 = address!("0x66efb28ac62686966ae85095ff3a772e014e7fbf56d4c5f6fac5606d4dde23a");
         let acc_2 = address!("0x6b86e40118f29ebe393a75469b4d926c7a44c2e2681b6d319520b7c1156d114");
@@ -1018,13 +832,6 @@ mod tests {
             state_root: felt!("0x99"),
             parent_hash: felt!("0x999"),
             gas_prices: GasPrices { eth: 1111, strk: 2222 },
-            // universal_deployer: Some(UniversalDeployerConfig {
-            //     class_hash: DEFAULT_LEGACY_UDC_CLASS_HASH,
-            //     address: address!(
-            //         "0x041a78e741e5af2fec34b695679bc6891742439f7afb8484ecd7766661ad02bf"
-            //     ),
-            //     storage: Some([(felt!("0x10"), felt!("0x100"))].into()),
-            // }),
         };
 
         assert_eq!(actual_genesis.number, expected_genesis.number);
@@ -1032,12 +839,6 @@ mod tests {
         assert_eq!(actual_genesis.timestamp, expected_genesis.timestamp);
         assert_eq!(actual_genesis.state_root, expected_genesis.state_root);
         assert_eq!(actual_genesis.gas_prices, expected_genesis.gas_prices);
-        // assert_eq!(actual_genesis.fee_token.address, expected_genesis.fee_token.address);
-        // assert_eq!(actual_genesis.fee_token.name, expected_genesis.fee_token.name);
-        // assert_eq!(actual_genesis.fee_token.symbol, expected_genesis.fee_token.symbol);
-        // assert_eq!(actual_genesis.fee_token.decimals, expected_genesis.fee_token.decimals);
-        // assert_eq!(actual_genesis.fee_token.class_hash, expected_genesis.fee_token.class_hash);
-        // assert_eq!(actual_genesis.universal_deployer, expected_genesis.universal_deployer);
         assert_eq!(actual_genesis.allocations.len(), expected_genesis.allocations.len());
 
         for alloc in actual_genesis.allocations {
@@ -1090,22 +891,6 @@ mod tests {
 
         let classes = BTreeMap::from([
             (
-                DEFAULT_LEGACY_UDC_CLASS_HASH,
-                GenesisClass {
-                    sierra: None,
-                    casm: DEFAULT_LEGACY_UDC_CASM.clone().into(),
-                    compiled_class_hash: DEFAULT_LEGACY_UDC_COMPILED_CLASS_HASH,
-                },
-            ),
-            (
-                DEFAULT_LEGACY_ERC20_CLASS_HASH,
-                GenesisClass {
-                    sierra: None,
-                    casm: DEFAULT_LEGACY_ERC20_CASM.clone().into(),
-                    compiled_class_hash: DEFAULT_LEGACY_ERC20_COMPILED_CLASS_HASH,
-                },
-            ),
-            (
                 DEFAULT_ACCOUNT_CLASS_HASH,
                 GenesisClass {
                     compiled_class_hash: DEFAULT_ACCOUNT_COMPILED_CLASS_HASH,
@@ -1124,15 +909,6 @@ mod tests {
             ),
         ]);
 
-        let fee_token = FeeTokenConfig {
-            address: DEFAULT_ETH_FEE_TOKEN_ADDRESS,
-            name: String::from("ETHER"),
-            symbol: String::from("ETH"),
-            decimals: 18,
-            class_hash: DEFAULT_LEGACY_ERC20_CLASS_HASH,
-            storage: None,
-        };
-
         let allocations = BTreeMap::from([(
             address!("0x66efb28ac62686966ae85095ff3a772e014e7fbf56d4c5f6fac5606d4dde23a"),
             GenesisAllocation::Account(GenesisAccountAlloc::Account(GenesisAccount {
@@ -1145,7 +921,6 @@ mod tests {
         )]);
 
         let expected_genesis = Genesis {
-            // fee_token,
             classes,
             allocations,
             number: 0,
@@ -1154,14 +929,8 @@ mod tests {
             parent_hash: felt!("0x999"),
             sequencer_address: address!("0x100"),
             gas_prices: GasPrices { eth: 1111, strk: 2222 },
-            // universal_deployer: Some(UniversalDeployerConfig {
-            //     class_hash: DEFAULT_LEGACY_UDC_CLASS_HASH,
-            //     address: DEFAULT_UDC_ADDRESS,
-            //     storage: None,
-            // }),
         };
 
-        // assert_eq!(actual_genesis.universal_deployer, expected_genesis.universal_deployer);
         assert_eq!(actual_genesis.allocations.len(), expected_genesis.allocations.len());
 
         for (address, alloc) in actual_genesis.allocations {
